rag/requirements.txt
flask
faiss-cpu
transformers
sentence-transformers
torch
